NAS-100746 All components except set prod
